@@ -1,10 +1,11 @@
 {
   "name": "sync_server_ts",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Sync Server written in TypeScript",
   "main": "main.js",
   "scripts": {
     "start": "npx ts-node src/main.ts",
+    "start:prod": "node ./dist/src/main.js",
     "build": "npx tsc",
     "lint": "eslint . --fix",
     "test": "echo \"Error: no test specified\" && exit 1"
@@ -12,26 +13,26 @@
   "author": "Stefano Falda",
   "license": "MIT",
   "devDependencies": {
-    "@tsconfig/node18": "^2.0.1",
-    "@types/express": "^4.17.17",
-    "@types/handlebars": "^4.1.0",
-    "@types/node": "^20.2.1",
-    "@types/nodemailer": "^6.4.8",
-    "@types/pg": "^8.6.6",
-    "@typescript-eslint/eslint-plugin": "^5.59.6",
-    "@typescript-eslint/parser": "^5.59.6",
-    "eslint": "^8.41.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.0.4"
+    "@tsconfig/node18": "^18.2.4",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/morgan": "^1.9.9",
+    "@types/node": "^22.5.4",
+    "@types/nodemailer": "^6.4.15",
+    "@types/pg": "^8.11.8",
+    "@typescript-eslint/eslint-plugin": "^8.4.0",
+    "@typescript-eslint/parser": "^8.4.0",
+    "eslint": "^9.10.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.4"
   },
   "dependencies": {
-    "aws-sdk": "^2.1397.0",
     "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "handlebars": "^4.7.7",
+    "express": "^4.19.2",
+    "handlebars": "^4.7.8",
     "morgan": "^1.10.0",
-    "nodemailer": "^6.9.3",
-    "pg": "^8.11.0",
-    "rotating-file-stream": "^3.1.0"
+    "nodemailer": "^6.9.15",
+    "pg": "^8.12.0",
+    "rotating-file-stream": "^3.2.3"
   }
 }
